Rename all snake_cased folders under src/components/elements/wellplates to camelCase
