+++ app/packs/src/components/elements/wellplates/WellplateDetails.js
@@ -11,11 +11,11 @@ import LoadingActions from 'src/stores/alt/actions/LoadingActions';
 import ElementCollectionLabels from 'src/components/elementsList/ElementCollectionLabels';
 import ElementActions from 'src/stores/alt/actions/ElementActions';
 import DetailActions from 'src/stores/alt/actions/DetailActions';
-import Wellplate from 'src/components/elements/wellplates/designer_tab/Wellplate';
-import WellplateList from 'src/components/elements/wellplates/list_tab/WellplateList';
-import WellplateProperties from 'src/components/elements/wellplates/properties_tab/WellplateProperties';
-import WellplateDetailsContainers from 'src/components/elements/wellplates/analyses_tab/WellplateDetailsContainers';
-import WellplateDetailsAttachments from 'src/components/elements/wellplates/attachments_tab/WellplateDetailsAttachments';
+import Wellplate from 'src/components/elements/wellplates/designerTab/Wellplate';
+import WellplateList from 'src/components/elements/wellplates/listTab/WellplateList';
+import WellplateProperties from 'src/components/elements/wellplates/propertiesTab/WellplateProperties';
+import WellplateDetailsContainers from 'src/components/elements/wellplates/analysesTab/WellplateDetailsContainers';
+import WellplateDetailsAttachments from 'src/components/elements/wellplates/attachmentsTab/WellplateDetailsAttachments';
 import PrintCodeButton from 'src/components/common/PrintCodeButton';
 import Attachment from 'src/models/Attachment';
 import Utils from 'src/utilities/Functions';
