+++ app/packs/src/components/elements/wellplates/analysesTab/WellplateDetailsContainers.js
